refactor: organize color-space functions into separate files
Move color utility functions into separate file.
