+++ src/css/index.ts
@@ -55,7 +55,8 @@ import {wordBreak} from './property-descriptors/word-break';
 import {zIndex} from './property-descriptors/z-index';
 import {CSSValue, isIdentToken, Parser} from './syntax/parser';
 import {Tokenizer} from './syntax/tokenizer';
-import {Color, color as colorType, isTransparent} from './types/color';
+import {Color, color as colorType} from './types/color';
+import {isTransparent} from './types/color-utilities';
 import {angle} from './types/angle';
 import {image} from './types/image';
 import {time} from './types/time';
