+++ src/dom/replaced-elements/iframe-element-container.ts
@@ -1,6 +1,7 @@
 import {ElementContainer} from '../element-container';
 import {parseTree} from '../node-parser';
-import {Color, parseColor, COLORS, isTransparent} from '../../css/types/color';
+import {Color, parseColor, COLORS} from '../../css/types/color';
+import {isTransparent} from '../../css/types/color-utilities';
 import {Context} from '../../core/context';
 
 export class IFrameElementContainer extends ElementContainer {
