+++ src/index.ts
@@ -1,5 +1,6 @@
 import {Bounds, parseBounds, parseDocumentSize} from './css/layout/bounds';
-import {COLORS, isTransparent, parseColor} from './css/types/color';
+import {COLORS, parseColor} from './css/types/color';
+import {isTransparent} from './css/types/color-utilities';
 import {CloneConfigurations, CloneOptions, DocumentCloner, WindowOptions} from './dom/document-cloner';
 import {isBodyElement, isHTMLElement, parseTree} from './dom/node-parser';
 import {CacheStorage} from './core/cache-storage';
