+++ src/render/canvas/foreignobject-renderer.ts
@@ -1,6 +1,6 @@
 import {RenderConfigurations} from './canvas-renderer';
 import {createForeignObjectSVG} from '../../core/features';
-import {asString} from '../../css/types/color';
+import {asString} from '../../css/types/color-utilities';
 import {Renderer} from '../renderer';
 import {Context} from '../../core/context';
